+++ Source~/CleverAdsSolutions.Common/Source/MediationExtras.cs
@@ -1,4 +1,4 @@
-//  Copyright © 2024 CAS.AI. All rights reserved.
+//  Copyright © 2025 CAS.AI. All rights reserved.
 
 using System;
 using System.Collections.Generic;